LU-7085 lov: trying smaller memory allocations
Reduce struct lov_io_sub to smaller memory usage
on wide-stripe file systems.

This is a back-ported patch from:
Lustre-commit: e1e56300cac30fe8d9db296107905f5936648c3c
Lustre-change: http://review.whamcloud.com/17476

Reviewed-on: http://review.whamcloud.com/18324
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/cl_object.h
@@ -2624,9 +2624,9 @@ void cl_sync_io_end(const struct lu_env *env, struct cl_sync_io *anchor);
  * \see lu_env, lu_context, lu_context_key
  * @{ */
 
-struct lu_env *cl_env_get        (int *refcheck);
-struct lu_env *cl_env_alloc      (int *refcheck, __u32 tags);
-void           cl_env_put        (struct lu_env *env, int *refcheck);
+struct lu_env *cl_env_get        (__u16 *refcheck);
+struct lu_env *cl_env_alloc      (__u16 *refcheck, __u32 tags);
+void           cl_env_put        (struct lu_env *env, __u16 *refcheck);
 unsigned       cl_env_cache_purge(unsigned nr);
 struct lu_env *cl_env_percpu_get (void);
 void           cl_env_percpu_put (struct lu_env *env);
