+++ lustre/llite/lcommon_cl.c
@@ -118,7 +118,7 @@ static struct lu_env *ccc_inode_fini_env = NULL;
  * pressure, when environments cannot be allocated.
  */
 static DEFINE_MUTEX(ccc_inode_fini_guard);
-static int dummy_refcheck;
+static __u16 dummy_refcheck;
 
 int ccc_global_init(struct lu_device_type *device_type)
 {
@@ -164,7 +164,7 @@ int cl_setattr_ost(struct inode *inode, const struct iattr *attr,
         struct lu_env *env;
         struct cl_io  *io;
         int            result;
-        int            refcheck;
+	__u16          refcheck;
 
         ENTRY;
 
@@ -232,7 +232,7 @@ int cl_file_inode_init(struct inode *inode, struct lustre_md *md)
                 }
         };
         int result = 0;
-        int refcheck;
+	__u16 refcheck;
 
 	LASSERT(md->body->mbo_valid & OBD_MD_FLID);
 	LASSERT(S_ISREG(inode->i_mode));
@@ -319,7 +319,7 @@ void cl_inode_fini(struct inode *inode)
 	struct lu_env           *env;
 	struct ll_inode_info    *lli  = ll_i2info(inode);
         struct cl_object        *clob = lli->lli_clob;
-        int refcheck;
+	__u16  refcheck;
         int emergency;
 
         if (clob != NULL) {
