+++ lustre/llite/lcommon_misc.c
@@ -133,7 +133,7 @@ int cl_get_grouplock(struct cl_object *obj, unsigned long gid, int nonblock,
         struct cl_lock         *lock;
         struct cl_lock_descr   *descr;
         __u32                   enqflags;
-        int                     refcheck;
+	__u16                   refcheck;
         int                     rc;
 
         env = cl_env_get(&refcheck);
