+++ lustre/osc/osc_lock.c
@@ -305,7 +305,7 @@ static int osc_lock_upcall(void *cookie, struct lustre_handle *lockh,
 	struct osc_lock         *oscl  = cookie;
 	struct cl_lock_slice    *slice = &oscl->ols_cl;
 	struct lu_env           *env;
-	int			refcheck;
+	__u16			refcheck;
 	int			rc;
 
 	ENTRY;
@@ -360,7 +360,7 @@ static int osc_lock_upcall_agl(void *cookie, struct lustre_handle *lockh,
 	struct osc_object	*osc = cookie;
 	struct ldlm_lock	*dlmlock;
 	struct lu_env           *env;
-	int			refcheck;
+	__u16			refcheck;
 	ENTRY;
 
 	env = cl_env_get(&refcheck);
@@ -394,7 +394,7 @@ static int osc_lock_flush(struct osc_object *obj, pgoff_t start, pgoff_t end,
 			  enum cl_lock_mode mode, int discard)
 {
 	struct lu_env		*env;
-	int			refcheck;
+	__u16			refcheck;
 	int			rc = 0;
 	int			rc2 = 0;
 
@@ -550,7 +550,7 @@ static int osc_ldlm_blocking_ast(struct ldlm_lock *dlmlock,
 	}
 	case LDLM_CB_CANCELING: {
 		struct lu_env *env;
-		int refcheck;
+		__u16 refcheck;
 
 		/*
 		 * This can be called in the context of outer IO, e.g.,
@@ -585,7 +585,7 @@ static int osc_ldlm_glimpse_ast(struct ldlm_lock *dlmlock, void *data)
 	struct lu_env		*env;
 	struct ost_lvb		*lvb;
 	struct req_capsule	*cap;
-	int			refcheck;
+	__u16			refcheck;
 	int			result;
 
 	ENTRY;
@@ -697,7 +697,7 @@ unsigned long osc_ldlm_weigh_ast(struct ldlm_lock *dlmlock)
 	struct osc_lock		*oscl;
 	unsigned long            weight;
 	bool			found = false;
-	int			refcheck;
+	__u16			refcheck;
 	ENTRY;
 
 	might_sleep();
